Removing authid requirement from LDAP Auth
The LDAP Module relayed in many parts on the correct content of "authid", which held the full DN of a user inside LDAP.

This will be changed to work with the default filters and settings specified in the XML configuration.

A admin should make sure that if multiple LDAP trees are used no overlapping in terms of the usernames exists.
(refs #3473)