Should fix PHP update to version > 5.3.3 crashing cronks
Fixed SplFileObject casting behaviour crashing PHP 5.3.3. AppKitFileUtil::getAlternateFilename
does not, as the method would suggest, return a filename, but an SplFileObject instead.
This is being cast to a string in file_exists() previously, but it seems that this behaviour
does not apply anymore in PHP 5.3.3+. So MetaInformationSuccessView now call $file->getRealPath()
to safely get the templates file path.

refs #3751