Fix CronksDataModel caching feature
We now add a timestamp to the cached data, to be able to determine if the cache
is still valid.

The timestamp is compared against the filemtime of the XML cache on disk
(cronks.xml cache).

In addition the handling of the param ($all) has been disabled in the
getXmlCronks function, it is handled in the function combinedData
(since 5da9fae0). This will avoid having only a partial cache.

fixes #3919