Fix race condition causing cronks to be destroyed
Cronk creation is wrapped in the Cronk.util.scriptInterface
class, which has it's own getParent() method. The issue was
that parentCmp was in the function closure and not in the object
scope. When the timing is bad, parentCmp could be overwritten
between the setup code of a Metagrid, causing a removeAll() call
to clear the wrong container.

fixes #3437