Revert "LongPluginOutput: Fix missing support of multi line in html"
This reverts commit 14fdf6323dde83ba51d61d29f7ff3ee4b23614f4.

The fix actually fixes nothing for the following reasons:

1) The "Long plugin output" field of a host's or service's detail window already has `\n' replaced with the HTML break tag.
This is because the information is gathered via Icinga Web's API which calls `Api_Result_OutputRewriteModel::rewritePluginOutput()'
in the controller `ApiSearchAction'.
Note that the carriage return `\r' is not taken into accout, but that's not a problem here.

2) Grids do not benefit from `Api_Result_OutputRewriteModel::rewritePluginOutput()' as its not being called. The `nl2br()' should've
helped here. According to http://www.ietf.org/rfc/rfc4627.txt the JSON representation of strings must escape control characters which includes the line feed `\n'
and the carriage return `\r'. So '\n', '\r' and '\r\n' become '\\n', '\\r' and '\\r\\n' respectively making them a textual representation of the control characters.
The `nl2br()' implementation however only replaces the unescaped line feed and carriage return which does of course replace nothing in the above case.

3) Introducing `String::nl2br()' while ExtJS provides `Ext.util.Format.nl2br()' is superfluous.

refs #2653